project-root/requirements.txt
discord.py==2.4.0
aiosqlite
python-dotenv
aiohttp
